pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
authors = [{ name = "Luís F. Pereira", email = "luisfgper@gmail.com" }]
name = "fmib"
dynamic = ["version"]
readme = "README.rst"
description = "Functional maps at Milano-Bicocca."
license = { file = "LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Programming Language :: Python :: 3",
]
requires-python = ">= 3.9"
dependencies = ["numpy", "scipy", "pyfmaps", "robust-laplacian", "libigl"]

[project.urls]
repository = "http://github.com/luisfpereira/fmib"

[tool.setuptools.dynamic]
version = { attr = "fmib.__version__" }

[tool.setuptools.packages.find]
include = ["fmib", "fmib.*"]
